+++ CHANGELOG.md
@@ -8,6 +8,10 @@ a release is made.
 
 ## [Unreleased]
 
+### Removed
+
+- The New scenario dialog no longer links to the Paint a Galaxy website.
+
 ## [0.9.0] - 2026-09-22
 
 ### Added